Added the ability to add > 150 character text
This is exceptionally buggy with, 1/9 requests responding with server error messages. In addition, things like sentences can be messed up due to them being cut short. Due to this a part of the context of the phrase is taken away and reduces the quality of the outputted text. As cool as this may be to do, it seems that the current method of making these requests may need to be changed.